BREAKING CHANGE: remove EnemyState, RefreshState; turns working, fix bugs

+++ spring-boot/src/main/java/org/springframework/ntfh/entity/user/UserService.java
@@ -15,6 +15,7 @@
 */
 package org.springframework.ntfh.entity.user;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -87,8 +88,9 @@ public class UserService {
 	}
 
 	@Transactional(readOnly = true)
-	public Page<User> findAllPage(Pageable pageable) {
-		return userRepository.findAllPage(pageable);
+	public List<User> findPage(Pageable pageable) {
+		Page<User> page = userRepository.findAllPage(pageable);
+		return page.getContent();
 	}
 
 	@Transactional(readOnly = true)
@@ -109,12 +111,12 @@ public class UserService {
 	/**
 	 * Update a user's information
 	 * 
+	 * @author andrsdt
 	 * @param user
 	 * @return
 	 * @throws DataAccessException
 	 * @throws NonMatchingTokenException
 	 * @throws DataIntegrityViolationException
-	 * @author andrsdt
 	 */
 	@Transactional
 	public User updateUser(User user, String token) throws DataAccessException, DataIntegrityViolationException,
@@ -132,34 +134,25 @@ public class UserService {
 
 		// Before updating, make sure there are no null values. If the user didn't send
 		// them in the form, they must stay the same as they were in the database.
-		User userInDatabase = this.findUser(user.getUsername());
-		if (user.getPassword() == null || user.getPassword().equals("null")) {
-			user.setPassword(userInDatabase.getPassword());
-		} else {
+
+		User userInDB = this.findUser(user.getUsername());
+		if (user.getEmail() != null) {
+			// If there is a new email, set it on the database
+			userInDB.setEmail(user.getEmail());
+		}
+		if (user.getPassword() != null) {
 			// If there is a new password input, encrypt it using bcrypt
 			String encodedParamPassword = passwordEncoder.encode(user.getPassword());
-			user.setPassword(encodedParamPassword);
+			userInDB.setPassword(encodedParamPassword);
 		}
-		if (user.getEmail() == null)
-			user.setEmail(userInDatabase.getEmail());
-		if (user.getEnabled() == null) {
-			user.setEnabled(userInDatabase.getEnabled());
-		}
-		return userRepository.save(user);
+		return userInDB;
 	}
 
 	@Transactional
 	public String loginUser(User user) throws DataAccessException, IllegalArgumentException, BannedUserException {
-		Optional<User> foundUserOptional = userRepository.findById(user.getUsername());
-		if (!foundUserOptional.isPresent()) {
-			// TODO move this validation to the findUser method in the service?
-			throw new DataAccessException("User not found") {
-			};
-		}
-
-		User userInDB = foundUserOptional.get();
+		User userInDB = this.findUser(user.getUsername());
 		if (!userInDB.getEnabled()) {
-			throw new BannedUserException("This user has been banned") {
+			throw new BannedUserException("You have been banned") {
 			};
 		}
 
@@ -172,31 +165,24 @@ public class UserService {
 
 	@Transactional
 	public User setCharacter(String username, Character character) throws DataAccessException {
-		Optional<User> foundUserOptional = userRepository.findById(username);
-		if (!foundUserOptional.isPresent()) {
-			throw new DataAccessException("User not found") {
-			};
-		}
-		User user = foundUserOptional.get();
-		user.setCharacter(character);
-		return user;
+		User userInDB = this.findUser(username);
+		userInDB.setCharacter(character);
+		return userInDB;
 
 	}
 
 	@Transactional
-	public User banUser(User user) throws DataAccessException {
-		Optional<User> foundUserOptional = userRepository.findById(user.getUsername());
-		if (!foundUserOptional.isPresent()) {
-			throw new DataAccessException("User not found") {
-			};
-		}
-		User userInDB = foundUserOptional.get();
-		userInDB.setEnabled(user.getEnabled());
-		return user;
+	public User toggleBanUser(String username, String token) throws DataAccessException {
+		User userInDB = this.findUser(username);
+		userInDB.setEnabled(!userInDB.getEnabled());
+		return userInDB;
 	}
 
 	@Transactional
 	public void deleteUser(User user) {
+		if (user.getLobby() != null && user.getLobby().getHasStarted()) {
+			throw new IllegalStateException("You cannot delete a user while he/she is playing a game");
+		}
 		this.userRepository.deleteById(user.getUsername());
 	}
